Loosen zoom definitions
There's only so much that can be done while pulling from the
pre-generalized tables which limit the features, but this wides
some of the zoom level selections.

Of course, once we start modifying the style, we may find these are
the wrong features to widen the zoom on, but there's not really a way
of knowing exactly what yet.